Add files via upload
Load the dataset as a Pandas data frame.
Display the first ten rows of data.
Find the dimensions (number of rows and columns) in the data frame. What do these two numbers represent in the context of the data?
Find the top five games by critic score.
Find the number of video games in the data frame in each genre.
Find the first five games in the data frame on the SNES platform.
Find the five publishers with the highest total global sales. Note: You will need to calculate the total global sales for each publisher to do this.
Create a new column in the data frame that calculates the percentage of global sales from North America. Display the first five rows of the new data frame.
Find the number NaN entries (missing data values) in each column.
Try to calculate the median user score of all the video games. You will likely run into an error because some of the user score entries are a non-numerical string that cannot be converted to a float. Find and replace this string with NaN and then calculate the median. Then, replace all NaN entries in the user score column with the median value.